fun.vangenuchten.se.h(soilwaterfun)	R Documentation

van Genuchten 1980's function for soil relative saturation.

Description

Calculate the relative saturation Se of a soil at a given tension h with the Van Genuchten water retention function.

Usage


fun.vangenuchten.se.h(h, alpha, n, cPar = 1)

Arguments

h 	 Vector of numerical. Matrix potential of the soil, in [m]. Values should be negative (suction).

alpha 	 Single numerical. alpha (shape) parameter of the Van Genuchten water retention function, in [m-1] (inverse length unit of h).

n 	 Single numerical. n shape parameter of the Van Genuchten water retention function, dimensionless [-]. See also the 'cPar' parameter that, along with 'n', is used to calculate van Genuchten's m parameter.

cPar 	 Single numerical. Value of the c parameter of the Van Genuchten water retention function, that allows to calculate the m parameter so m = (1 - cPar/n). Dimensionless [-]. Usually fixed / constant.

Value

The function returns the relative water content (degree of saturation, Se, [-]).

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

References

van Genuchten M. Th., 1980. A closed form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44:892-898. Kutilek M. & Nielsen D.R., 1994. Soil hydrology. Catena-Verlag, GeoEcology textbook, Germany. ISBN: 9-923381-26-3., 370 p.

Examples

require( "soilwaterfun" ) 

# Example with the properties of the Footprint soil type P22i, 
# 3rd layer:
fun.vangenuchten.se.h( 
    h       = -c(0,0.01,0.1,1,10,100,158), 
    alpha   = 3.561099, 
    n       = 1.212074  
)   #

[Package soilwaterfun version 1.0 Index]