fun.vangenuchten.theta.h(soilwaterfun)	R Documentation

van Genuchten 1980's function theta(h) (water retension).

Description

Calculate the water content theta at a given tension h with the Van Genuchten water retention function.

Usage


fun.vangenuchten.theta.h(h, alpha, n, cPar = 1, thetaS, thetaR)

Arguments

h 	 Vector of numerical. Matrix potential of the soil, in [m]. Values should be negative (suction).

alpha 	 Single numerical. alpha (shape) parameter of the Van Genuchten water retention function, in [m-1] (inverse length unit of h).

n 	 Single numerical. n shape parameter of the Van Genuchten water retention function, dimensionless [-]. See also the 'cPar' parameter that, along with 'n', is used to calculate van Genuchten's m parameter.

cPar 	 Single numerical. Value of the c parameter of the Van Genuchten water retention function, that allows to calculate the m parameter so m = (1 - cPar/n). Dimensionless [-].

thetaS 	 Single numerical. Saturated water content of the soil, in [-] or [m3 of water.m-3 of bulk soil].

thetaR 	 Single numerical. Residual water content of the soil, in [-] or [m3 of water.m-3 of bulk soil].

Value

The function returns the water content [m3.m-3] at the given tension h.

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

References

van Genuchten M. Th., 1980. A closed form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44:892-898. Kutilek M. & Nielsen D.R., 1994. Soil hydrology. Catena-Verlag, GeoEcology textbook, Germany. ISBN: 9-923381-26-3., 370 p.

Examples

require( "soilwaterfun" ) 

# Example with the properties of the Footprint soil type P22i, 
# 3rd layer:
fun.vangenuchten.theta.h( 
    h       = -c(0,0.01,0.1,1,10,100,158), 
    alpha   = 3.561099, 
    n       = 1.212074, 
    thetaS  = 0.4162380, 
    thetaR  = 0  
)   #

curve( 
    fun.vangenuchten.theta.h( 
        h       = -x, 
        alpha   = 3.561099, 
        n       = 1.212074, 
        thetaS  = 0.4162380, 
        thetaR  = 0  
    ),  #
    xlim = c(0.001,158), 
    col  = "red", 
    log  = "x", 
    xlab = "-h", 
    ylab = expression( theta ) 
)   #

[Package soilwaterfun version 1.0 Index]