soilwaterfun-package(soilwaterfun)	R Documentation

Functions for soil water retention and soil hydraulic conductivity.

Description

The package implements commonly used functions for soil water retention (water content = f(matrix potential)) and soil hydraulic conductivity (hydraulic conductivity K = f(matrix potential), or = f(water content)). The functions for for soil water retention implemented are: van Genuchten 1980; Brooks & Corey 1964; Campbell 1974. The functions for soil hydraulic conductivity implemented are: Mualem 1976; Brooks & Corey 1964; Campbell 1974. NB: The package soilwaterptf contains pedotransfer function for estimating the parameters of these functions from basic soil properties (valid for european soils).

Details

Package: 	 soilwaterfun

Version: 	 1.0

Date: 	 2010-09-20

Title: 	 Functions for soil water retention and soil hydraulic conductivity.

Author: 	 Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

Maintainer: 	 Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

LazyLoad: 	 yes

Depends: 	 R (>= 2.4.1), MASS

Suggests: 	 soilwaterptf

License: 	 AGPL-3

URL: 	 http://www.mark.slu.se, http://ckb.slu.se, http://julienmoeys.free.fr

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soilwaterfun version 1.0 Index]