soilwaterptf-package(soilwaterptf)	R Documentation

Wosten et al. 1999 pedotransfer functions for soil hydraulic properties.

Description

Pedotransfer functions (PTFs) for estimating the physical parameters of soil water retention and hydraulic conductivity functions from easily available soil properties (typically soil texture, organic matter content, bulk density). The pedotransfer function(s) implemeneted is (are): (1) Wosten et al. 1999 (Geoderma 90:169-185) pedotransfer functions for estimating the parameters of van-Genuchten (1980) Mualem (1976) water retention and hydraulic conductivity functions (valid for European soils).

Details

Package: 	 soilwaterptf

Version: 	 1.01

Date: 	 2010-09-20

Title: 	 Wosten et al. 1999 pedotransfer functions for soil hydraulic properties.

Author: 	 Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

Maintainer: 	 Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

LazyLoad: 	 yes

Depends: 	 R (>= 2.4.1), MASS

Suggests: 	 soilwaterfun

License: 	 AGPL-3

URL: 	 http://www.mark.slu.se, http://ckb.slu.se, http://julienmoeys.free.fr

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soilwaterptf version 1.01 Index]